requirements.txt
# This is the requirements.txt file in the shoplifting_detection\ directory
torch
torchvision
transformers
opencv-python
numpy
pyyaml
tensorboard
tqdm
Pillow
easydict